Add files via upload
boyer_moore.c – The main C program file that implements the Boyer–Moore string matching algorithm, reads input files, and displays the results.
data.txt – The dataset file containing the text of “The Adventures of Sherlock Holmes” downloaded from Project Gutenberg. It serves as the main text in which the pattern is searched.
pattern.txt – A small text file containing the word or phrase (pattern) to be searched in the dataset.